Updated to not use bootstrap & be more responsive
* Updated the index page so the banner looks nice on mobile and now
doesn't break with the github and patreon buttons om iPhone (Still not
quite sure why it was). Apparently the banner now doesn't break, but
it's fine for now so will update on that next time...

* Updated the commands page to have the categories as a menu on the side
rather than tabs on the top. Changed it all around so that it actually
looks good on mobile.

* The FAQs and changelog pages got almost no changes, but pretty much
just look the same I think... They look good on mobile sized window at
least, so that's good. Kinda forgot to check the before, and don't feel
like reverting stuff to do so